Description: fix arg to printf() to avoid build error
git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/private-branches/fab4/squeezeplay@5739 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9

+++ src/squeezeplay/src/log.c
@@ -76,7 +76,7 @@ void log_category_vlog(struct log_category *category, enum log_priority priority
 		printf("%04d%02d%02d %02d:%02d:%02d.%03ld %-6s %s - %s\n",
 		       tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday,
 		       tm.tm_hour, tm.tm_min, tm.tm_sec,
-		       t.tv_usec / 1000,
+		      (long)(t.tv_usec / 1000),
 		       log_priority_to_string(priority), category->name, buf);
 	}
